Update sentence_translations.json
Hi! I am a native Russian with some mathematical background (MsC in mathematics). I love this channel, and I hope this project will help share it with a broader audience around the world. As a small contribution, I improved subtitles for this video and hope they are good enough now. Please note that mathematical letters x and y in russian are inferred from latin and are pronounced as "икс" and "игрек", Euler constant e is pronounced as the same russian letter "е", while s would be just english s:) This is a little messy and might complicate reading mathematical notations, but hopefully in the current form it will be fine.